Rename Data to Types.
We intend to use the namespace Data to hold data used to drive
computations (e.g. values fixed by law to compute taxes or salaries).

We'll use Types as a namespace for the pure code (as opposed to STM or
IO layers).
