+++ src/Curiosity/Command.hs
@@ -20,15 +20,15 @@ module Curiosity.Command
 import qualified Data.String as Str
 import Web.HttpApiData
 import qualified Curiosity.Data                as Data
-import qualified Curiosity.Data.Business       as Business
-import qualified Curiosity.Data.Email          as Email
-import qualified Curiosity.Data.Employment     as Employment
-import qualified Curiosity.Data.Invoice        as Invoice
-import qualified Curiosity.Data.Legal          as Legal
-import qualified Curiosity.Data.Order          as Order
-import qualified Curiosity.Data.Quotation      as Quotation
-import qualified Curiosity.Data.SimpleContract as SimpleContract
-import qualified Curiosity.Data.User           as User
+import qualified Curiosity.Types.Business       as Business
+import qualified Curiosity.Types.Email          as Email
+import qualified Curiosity.Types.Employment     as Employment
+import qualified Curiosity.Types.Invoice        as Invoice
+import qualified Curiosity.Types.Legal          as Legal
+import qualified Curiosity.Types.Order          as Order
+import qualified Curiosity.Types.Quotation      as Quotation
+import qualified Curiosity.Types.SimpleContract as SimpleContract
+import qualified Curiosity.Types.User           as User
 import qualified Curiosity.Parse               as P
 import qualified Data.Text                     as T
 import qualified Options.Applicative           as A
