+++ src/Curiosity/Html/Action.hs
@@ -12,8 +12,8 @@ module Curiosity.Html.Action
   , EchoPage'(..)
   ) where
 
-import qualified Curiosity.Data.Quotation      as Quotation
-import qualified Curiosity.Data.User           as User
+import qualified Curiosity.Types.Quotation     as Quotation
+import qualified Curiosity.Types.User          as User
 import           Curiosity.Html.Misc
 import qualified Smart.Html.Dsl                as Dsl
 import           Smart.Html.Panel               ( Panel(..) )
