+++ src/Curiosity/Html/Invoice.hs
@@ -7,8 +7,8 @@ module Curiosity.Html.Invoice
   , CreateInvoicePage(..)
   ) where
 
-import qualified Curiosity.Data.Invoice        as Invoice
-import qualified Curiosity.Data.User           as User
+import qualified Curiosity.Types.Invoice       as Invoice
+import qualified Curiosity.Types.User          as User
 import           Curiosity.Html.Misc
 import qualified Text.Blaze.Html5              as H
 import           Text.Blaze.Html5               ( (!) )
