+++ src/Curiosity/Html/Misc.hs
@@ -51,7 +51,7 @@ module Curiosity.Html.Misc
   , autoReload
   ) where
 
-import qualified Curiosity.Data.User           as User
+import qualified Curiosity.Types.User          as User
 import           Curiosity.Html.Navbar          ( navbar
                                                 , navbarWebsite
                                                 )
