+++ src/Curiosity/Html/SimpleContract.hs
@@ -19,9 +19,9 @@ module Curiosity.Html.SimpleContract
   , ConfirmSimpleContractPage(..)
   ) where
 
-import qualified Curiosity.Data.Country        as Country
-import qualified Curiosity.Data.SimpleContract as SimpleContract
-import qualified Curiosity.Data.User           as User
+import qualified Curiosity.Types.Country        as Country
+import qualified Curiosity.Types.SimpleContract as SimpleContract
+import qualified Curiosity.Types.User           as User
 import           Curiosity.Html.Misc
 import qualified Smart.Html.Alert              as Alert
 import qualified Smart.Html.Button             as Button
