+++ src/Curiosity/Interpret.hs
@@ -37,7 +37,7 @@ module Curiosity.Interpret
 
 import qualified Curiosity.Command             as Command
 import qualified Curiosity.Data                as Data
-import qualified Curiosity.Data.User           as User
+import qualified Curiosity.Types.User          as User
 import qualified Curiosity.Parse               as P
 import qualified Curiosity.Runtime             as Rt
 import           Data.List                      ( isInfixOf, last )
