+++ src/Curiosity/Types/Business.hs
@@ -1,10 +1,10 @@
 {-# LANGUAGE DataKinds #-}
 {-# LANGUAGE DeriveAnyClass #-}
 {- |
-Module: Curiosity.Data.Business
+Module: Curiosity.Types.Business
 Description: Business entities related datatypes
 -}
-module Curiosity.Data.Business
+module Curiosity.Types.Business
   ( Unit(..)
   , Create(..)
   , Update(..)
@@ -16,8 +16,8 @@ module Curiosity.Data.Business
   ) where
 
 import qualified Commence.Types.Wrapped        as W
-import qualified Curiosity.Data.PrefixedId     as Pre
-import qualified Curiosity.Data.User           as User
+import qualified Curiosity.Types.PrefixedId    as Pre
+import qualified Curiosity.Types.User          as User
 import           Data.Aeson
 import qualified Text.Blaze.Html5              as H
 import           Web.FormUrlEncoded             ( FromForm(..)
